# docs/swagger.yaml
basePath: /api
definitions:
  domain.Client:
    properties:
      company_id:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
      settings:
        $ref: '#/definitions/domain.ClientSettings'
      vehicles:
        items:
          $ref: '#/definitions/domain.ClientVehicle'
        type: array
    type: object
  domain.ClientSettings:
    properties:
      discount:
        type: integer
    type: object
  domain.ClientVehicle:
    properties:
      distance:
        type: integer
      id:
        type: string
      make:
        type: string
      model:
        type: string
      plate_number:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
...
  http.clientCreateRequest:
    properties:
      name:
        type: string
      phone_number:
        type: string
      settings:
        $ref: '#/definitions/domain.ClientSettings'
      vehicles:
        items:
          $ref: '#/definitions/domain.ClientVehicle'
        type: array
    type: object
  http.clientUpdateRequest:
    properties:
      name:
        type: string
      phone_number:
        type: string
      settings:
        $ref: '#/definitions/domain.ClientSettings'
      vehicles:
        items:
          $ref: '#/definitions/domain.ClientVehicle'
        type: array
    type: object
  http.clientVehicleCreateRequest:
    properties:
      distance:
        type: integer
      make:
        type: string
      model:
        type: string
      plate_number:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
  http.clientVehicleUpdateRequest:
    properties:
      distance:
        type: integer
      make:
        type: string
      model:
        type: string
      plate_number:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
...
  http.idResponse:
    properties:
      id:
        type: string
    type: object
  http.response:
    properties:
      message:
        type: string
    type: object
...
host: localhost:80
info:
  contact: {}
  title: Syncrepair API
  version: "1.0"
paths:
  /clients:
    get:
      consumes:
      - application/json
      description: Get all clients
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Client'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Get all
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: Create client
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.clientCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Create
      tags:
      - clients
  /clients/{clientID}/vehicles:
    get:
      consumes:
      - application/json
      description: Get all client's vehicles
      parameters:
      - description: Client ID
        in: path
        name: clientID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ClientVehicle'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Get all
      tags:
      - clients
    post:
      consumes:
      - application/json
      description: Create vehicle for client
      parameters:
      - description: Client ID
        in: path
        name: clientID
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.clientVehicleCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.idResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Create vehicle
      tags:
      - clients
  /clients/{clientID}/vehicles/{vehicleID}:
    delete:
      consumes:
      - application/json
      description: Delete client's vehicle by ID
      parameters:
      - description: Client ID
        in: path
        name: clientID
        required: true
        type: string
      - description: Vehicle ID
        in: path
        name: vehicleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Delete vehicle by ID
      tags:
      - clients
    get:
      consumes:
      - application/json
      description: Get client's vehicle by ID
      parameters:
      - description: Client ID
        in: path
        name: clientID
        required: true
        type: string
      - description: Vehicle ID
        in: path
        name: vehicleID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.ClientVehicle'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Get by ID
      tags:
      - clients
    put:
      consumes:
      - application/json
      description: Update client's vehicle by ID
      parameters:
      - description: Client ID
        in: path
        name: clientID
        required: true
        type: string
      - description: Vehicle ID
        in: path
        name: vehicleID
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.clientVehicleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Update vehicle by ID
      tags:
      - clients
  /clients/{id}:
    delete:
      consumes:
      - application/json
      description: Delete client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Delete
      tags:
      - clients
    get:
      consumes:
      - application/json
      description: Get client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Client'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Get by ID
      tags:
      - clients
    put:
      consumes:
      - application/json
      description: Update client by ID
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.clientUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.response'
      security:
      - UserAuth: []
      summary: Update
      tags:
      - clients
...
securityDefinitions:
  UserAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"